requirements.txt
flask
librosa
numpy
scikit-learn
joblib
gunicorn
resampy
tensorflow-cpu
Pillow